zai-engine/requirements-working-v1.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
python-multipart==0.0.9

# Database and Search
motor==3.3.2
pymongo==4.6.1
requests>=2.31.0
aiohttp==3.9.1
elastic-transport[async]==8.15.1
elasticsearch[async]==8.17.0
elasticsearch-dsl==8.11.0

# RAG Stack
haystack-ai==2.8.0
haystack-core-integrations[elasticsearch]>=0.1.0,<0.2.0
sentence-transformers>=3.0.0
huggingface-hub>=0.20.0
docling==2.14.0

# LLM Integration
openai==1.59.3
langchain-core>=0.1.0,<0.2.0
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-google-genai==0.0.3
google-generativeai>=0.3.1,<0.4.0

# Document Processing
python-magic==0.4.27
python-docx==1.1.2
PyPDF2==3.0.1
pytesseract==0.3.10

# Utils
tenacity==8.2.3
loguru==0.7.2
numpy==1.26.2
PyYAML==6.0.1
